+++ stdnet/contrib/djstdnet/tests/__init__.py
@@ -1,0 +1,4 @@
+try:
+    from test_djstdnet import *
+except ImportError:
+    pass