+++ stdnet/contrib/djstdnet/tests/settings.py
@@ -1,0 +1,31 @@
+DATABASES = {
+    'default': {
+        'ENGINE': 'django.db.backends.sqlite3',
+        'NAME': 'test-djstdnet.sqlite'
+    }
+}
+
+
+INSTALLED_APPS  = ['django.contrib.auth',
+                   'django.contrib.sessions',
+                   'django.contrib.sites',
+                   'django.contrib.contenttypes',
+                   'django.contrib.admin',
+                   #'djpcms',
+                   'stdnet.contrib.djstdnet'
+                   'stdnet.contrib.djstdnet.tests.testmodel']
+
+# Silence logging
+import logging
+
+class Silence(logging.Handler):
+    def emit(self, record):
+        pass
+
+logging.getLogger("djstdnet").addHandler(Silence())
+
+TEMPLATE_CONTEXT_PROCESSORS = (
+            "django.contrib.auth.context_processors.auth",
+            "django.contrib.messages.context_processors.messages",
+            "djpcms.core.context_processors.djpcms"
+            )