+++ stdnet/tests/__init__.py
@@ -18,3 +18,4 @@ from testme import *
 
 # Contrib
 from stdnet.contrib.timeserie.tests import *
+from stdnet.contrib.djstdnet.tests import *